refactor: fix len_zero, let_and_return, strip_prefix, match_like_matches, needless_return, redundant_pattern, useless_conversion clippy warnings
Applied via cargo clippy --fix across 77 files.

+++ src/storage/engines/impls/sst/trait_impl.rs
@@ -153,7 +153,7 @@ impl UnifiedStorageEngine for SstEngine {
         if let Ok(entries) = fs.list(&storage_url).await {
             for entry in &entries {
                 if !entry.metadata.is_directory {
-                    total_bytes += entry.metadata.size as u64;
+                    total_bytes += entry.metadata.size;
                     if entry.url.ends_with(".sst") || entry.url.ends_with(".proximablock") {
                         file_count += 1;
                     }
